Добавляет файлы для модулей
