+++ index.html
@@ -231,6 +231,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+  <script src="main.js" type="mosule"></script>
 
 </body>
 </html>
